Make clientId mapping selectors private
Move getInternalClientId and getExternalClientId from public selectors
to private-selectors.js since they are internal implementation details
for entity navigation.
